+++ lib/td_dd/access/policy.ex
@@ -3,6 +3,6 @@ defmodule TdDd.Access.Policy do
 
   @behaviour Bodyguard.Policy
 
-  def authorize(_action, %{role: role}, _params), do: role in ["admin", "service"]
+  def authorize(_action, %{role: role}, _params) when role in ["admin", "service"], do: true
   def authorize(_action, _claims, _params), do: false
 end