+++ lib/td_dq_web/views/search_view.ex
@@ -17,11 +17,15 @@ defmodule TdDqWeb.SearchView do
         Map.put(acc, :user_permissions, user_permissions)
 
       {:rules, rules}, acc ->
-        data = render_many(rules, RuleView, "rule.json")
+        data = render_many(rules, RuleView, "rule.json", %{lang: Map.get(assigns, :locale)})
         Map.put(acc, :data, data)
 
       {:implementations, implementations}, acc ->
-        data = render_many(implementations, ImplementationView, "implementation.json")
+        data =
+          render_many(implementations, ImplementationView, "implementation.json", %{
+            lang: Map.get(assigns, :locale)
+          })
+
         Map.put(acc, :data, data)
 
       {:scroll_id, scroll_id}, acc ->